fix bug where clients would only contact one replica
This bug exists because our setup is different from the original gryff paper. They had physical nodes with aritificial latency for each group of clients. In our setup, all clients are simulated on a single machine, so the latency between the client machine and the replica machines are the same for all clients.